♻️ move some code from `OpType` header to a dedicated cpp file
